linkcheck builder: enable streaming mode (no-immediate-content-read) during HTTP HEAD requests
This shouldn't theoretically make much of a difference for HTTP HEAD requests, since they're not expected to include body content in the response data.  In practice, I think this may resolve some test flakiness for sphinx.

Enabling streaming mode has to be done carefully to ensure that the connection is returned to the pool; we are using a context-manager when making requests here, so this should be safe.  Ref: https://requests.readthedocs.io/en/latest/user/advanced/#body-content-workflow